add authorization rules

+++ src/backend/api/Fusion.Resources.Api/Controllers/Requests/TasksController.cs
@@ -1,6 +1,9 @@
-﻿using Fusion.Resources.Domain;
+﻿using Fusion.AspNetCore.FluentAuthorization;
+using Fusion.Authorization;
+using Fusion.Resources.Domain;
 using Fusion.Resources.Domain.Commands;
 using Fusion.Resources.Domain.Commands.Tasks;
+using Fusion.Resources.Domain.Queries;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
@@ -18,6 +21,39 @@ namespace Fusion.Resources.Api.Controllers.Requests
         [HttpPost("requests/{requestId}/tasks")]
         public async Task<ActionResult> AddRequestTask([FromRoute] Guid requestId, [FromBody] AddRequestTaskRequest request)
         {
+            var requestItem = await DispatchAsync(new GetResourceAllocationRequestItem(requestId));
+            if (requestItem is null) return FusionApiError.NotFound(requestId, $"Request with id '{requestId}' was not found.");
+
+            #region Authorization
+            var authResult = await Request.RequireAuthorizationAsync(r =>
+            {
+                r.AlwaysAccessWhen().FullControl().FullControlInternal();
+                r.AnyOf(or =>
+                {
+                    if (requestItem.OrgPositionId.HasValue)
+                        or.OrgChartPositionWriteAccess(requestItem.Project.OrgProjectId, requestItem.OrgPositionId.Value);
+
+                    if (requestItem.AssignedDepartment is not null)
+                    {
+                        or.BeResourceOwner(
+                            new DepartmentPath(requestItem.AssignedDepartment).GoToLevel(2),
+                            includeParents: false,
+                            includeDescendants: true
+                        );
+                    }
+                    else
+                    {
+                        or.BeResourceOwner();
+                    }
+
+                    or.BeRequestCreator(requestId);
+                });
+            });
+
+            if (authResult.Unauthorized)
+                return authResult.CreateForbiddenResponse();
+            #endregion
+
             var command = new AddRequestTask(requestId, request.Title, request.Body, request.Category, request.Type)
             {
                 SubType = request.SubType,
@@ -45,6 +81,39 @@ namespace Fusion.Resources.Api.Controllers.Requests
         [HttpGet("requests/{requestId}/tasks")]
         public async Task<ActionResult> GetRequestTasks([FromRoute] Guid requestId)
         {
+            var request = await DispatchAsync(new GetResourceAllocationRequestItem(requestId));
+            if (request is null) return FusionApiError.NotFound(requestId, $"Request with id '{requestId}' was not found.");
+
+            #region Authorization
+            var authResult = await Request.RequireAuthorizationAsync(r =>
+            {
+                r.AlwaysAccessWhen().FullControl().FullControlInternal();
+                r.AnyOf(or =>
+                {
+                    if (request.OrgPositionId.HasValue)
+                        or.OrgChartPositionWriteAccess(request.Project.OrgProjectId, request.OrgPositionId.Value);
+
+                    if (request.AssignedDepartment is not null)
+                    {
+                        or.BeResourceOwner(
+                            new DepartmentPath(request.AssignedDepartment).GoToLevel(2),
+                            includeParents: false,
+                            includeDescendants: true
+                        );
+                    }
+                    else
+                    {
+                        or.BeResourceOwner();
+                    }
+
+                    or.BeRequestCreator(requestId);
+                });
+            });
+
+            if (authResult.Unauthorized)
+                return authResult.CreateForbiddenResponse();
+            #endregion
+
             var command = new GetRequestTasks(requestId);
             var tasks = await DispatchAsync(command);
 
@@ -54,9 +123,41 @@ namespace Fusion.Resources.Api.Controllers.Requests
         [HttpGet("requests/{requestId}/tasks/{taskId}")]
         public async Task<ActionResult> GetRequestTask([FromRoute] Guid requestId, [FromRoute] Guid taskId)
         {
+            var request = await DispatchAsync(new GetResourceAllocationRequestItem(requestId));
+            if (request is null) return FusionApiError.NotFound(requestId, $"Request with id '{requestId}' was not found.");
+
+            #region Authorization
+            var authResult = await Request.RequireAuthorizationAsync(r =>
+            {
+                r.AlwaysAccessWhen().FullControl().FullControlInternal();
+                r.AnyOf(or =>
+                {
+                    if (request.OrgPositionId.HasValue)
+                        or.OrgChartPositionWriteAccess(request.Project.OrgProjectId, request.OrgPositionId.Value);
+
+                    if (request.AssignedDepartment is not null)
+                    {
+                        or.BeResourceOwner(
+                            new DepartmentPath(request.AssignedDepartment).GoToLevel(2),
+                            includeParents: false,
+                            includeDescendants: true
+                        );
+                    }
+                    else
+                    {
+                        or.BeResourceOwner();
+                    }
+
+                    or.BeRequestCreator(requestId);
+                });
+            });
+
+            if (authResult.Unauthorized)
+                return authResult.CreateForbiddenResponse();
+            #endregion
+
             var command = new GetRequestTask(requestId, taskId);
             var task = await DispatchAsync(command);
-
             if (task is null) return FusionApiError.NotFound(taskId, $"A task with id '{taskId}' was not found on request with id '{requestId}'.");
 
             return Ok(new ApiRequestTask(task));
@@ -65,6 +166,39 @@ namespace Fusion.Resources.Api.Controllers.Requests
         [HttpPatch("requests/{requestId}/tasks/{taskId}")]
         public async Task<ActionResult> UpdateRequestTask([FromRoute] Guid requestId, [FromRoute] Guid taskId, [FromBody] PatchRequestTaskRequest patch)
         {
+            var request = await DispatchAsync(new GetResourceAllocationRequestItem(requestId));
+            if (request is null) return FusionApiError.NotFound(requestId, $"Request with id '{requestId}' was not found.");
+
+            #region Authorization
+            var authResult = await Request.RequireAuthorizationAsync(r =>
+            {
+                r.AlwaysAccessWhen().FullControl().FullControlInternal();
+                r.AnyOf(or =>
+                {
+                    if (request.OrgPositionId.HasValue)
+                        or.OrgChartPositionWriteAccess(request.Project.OrgProjectId, request.OrgPositionId.Value);
+
+                    if (request.AssignedDepartment is not null)
+                    {
+                        or.BeResourceOwner(
+                            new DepartmentPath(request.AssignedDepartment).GoToLevel(2),
+                            includeParents: false,
+                            includeDescendants: true
+                        );
+                    }
+                    else
+                    {
+                        or.BeResourceOwner();
+                    }
+
+                    or.BeRequestCreator(requestId);
+                });
+            });
+
+            if (authResult.Unauthorized)
+                return authResult.CreateForbiddenResponse();
+            #endregion
+
             var command = new UpdateRequestTask(requestId, taskId);
 
             if (patch.Title.HasValue) command.Title = patch.Title.Value;
@@ -89,6 +223,39 @@ namespace Fusion.Resources.Api.Controllers.Requests
         [HttpDelete("requests/{requestId}/tasks/{taskId}")]
         public async Task<ActionResult> DeleteRequestTask([FromRoute] Guid requestId, [FromRoute] Guid taskId)
         {
+            var request = await DispatchAsync(new GetResourceAllocationRequestItem(requestId));
+            if (request is null) return FusionApiError.NotFound(requestId, $"Request with id '{requestId}' was not found.");
+
+            #region Authorization
+            var authResult = await Request.RequireAuthorizationAsync(r =>
+            {
+                r.AlwaysAccessWhen().FullControl().FullControlInternal();
+                r.AnyOf(or =>
+                {
+                    if (request.OrgPositionId.HasValue)
+                        or.OrgChartPositionWriteAccess(request.Project.OrgProjectId, request.OrgPositionId.Value);
+
+                    if (request.AssignedDepartment is not null)
+                    {
+                        or.BeResourceOwner(
+                            new DepartmentPath(request.AssignedDepartment).GoToLevel(2),
+                            includeParents: false,
+                            includeDescendants: true
+                        );
+                    }
+                    else
+                    {
+                        or.BeResourceOwner();
+                    }
+
+                    or.BeRequestCreator(requestId);
+                });
+            });
+
+            if (authResult.Unauthorized)
+                return authResult.CreateForbiddenResponse();
+            #endregion
+
             var command = new DeleteRequestTask(requestId, taskId);
             var wasDeleted = await DispatchAsync(command);
